# PromQL: short-circuit vector binary ops