@bboreham 2021-09-20
Like this?  https://github.com/prometheus/prometheus/blob/03d084f8629477907cab39fc3d314b375eeac010/promql/engine.go#L1086

At that particular point I think the inputs are freshly-allocated in `vectors` earlier, and `vectors` does not get used again.
But it would probably be worth setting rules on how those values may be treated.